task: adding sample image logo
