+++ internal/sample/home.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <title>Sample: Home Page</title>
+    <script src="https://cdn.tailwindcss.com"></script>
+  </head>
+  <body class="bg-gray-50">
+    <main class="max-w-[1200px] mx-auto pt-10 text-center">
+        <h1 class="text-3xl font-bold">Public Page</h1>
+        <p class="mb-10">This is the public space, if you want to go to the secure page you need to login.</p>
+        <a href="/auth/login" class="text-white py-3 px-4 rounded-full bg-blue-400">Login</a>
+    </main>
+  </body>
+</html>
