+++ internal/sample/private.go
@@ -1,7 +1,20 @@
 package sample
 
-import "net/http"
+import (
+	_ "embed"
+	"net/http"
+)
 
-func Home(w http.ResponseWriter, r *http.Request) {
-	w.Write([]byte("Public section"))
+//go:embed private.html
+var private []byte
+
+// Private handler to show the private content to the user
+func Private(w http.ResponseWriter, r *http.Request) {
+	_, err := r.Cookie("sample")
+	if err != nil {
+		http.Redirect(w, r, "/auth/login", http.StatusFound)
+		return
+	}
+
+	w.Write(private)
 }
