+++ internal/sample/private.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <title>Sample: Home Page</title>
+    <script src="https://cdn.tailwindcss.com"></script>
+  </head>
+  <body class="bg-gray-50">
+    <main class="max-w-[1200px] mx-auto pt-10">
+        <h1 class="text-3xl font-bold">🔒 Private Section</h1>
+        <p>This is the private space, means you were able to login!</p>
+        <a href="/auth/login" class="text-blue-500 underline">Login</a>
+    </main>
+  </body>
+</html>
